Move path_to_url() to utils/urls.py.
