+++ tests/unit/test_urls.py
@@ -1,9 +1,10 @@
+import os
 import sys
 
 import pytest
+from pip._vendor.six.moves.urllib import request as urllib_request
 
-from pip._internal.utils.misc import path_to_url
-from pip._internal.utils.urls import get_url_scheme, url_to_path
+from pip._internal.utils.urls import get_url_scheme, path_to_url, url_to_path
 
 
 @pytest.mark.parametrize("url,expected", [
@@ -14,6 +15,22 @@ from pip._internal.utils.urls import get_url_scheme, url_to_path
 ])
 def test_get_url_scheme(url, expected):
     assert get_url_scheme(url) == expected
+
+
+@pytest.mark.skipif("sys.platform == 'win32'")
+def test_path_to_url_unix():
+    assert path_to_url('/tmp/file') == 'file:///tmp/file'
+    path = os.path.join(os.getcwd(), 'file')
+    assert path_to_url('file') == 'file://' + urllib_request.pathname2url(path)
+
+
+@pytest.mark.skipif("sys.platform != 'win32'")
+def test_path_to_url_win():
+    assert path_to_url('c:/tmp/file') == 'file:///C:/tmp/file'
+    assert path_to_url('c:\\tmp\\file') == 'file:///C:/tmp/file'
+    assert path_to_url(r'\\unc\as\path') == 'file://unc/as/path'
+    path = os.path.join(os.getcwd(), 'file')
+    assert path_to_url('file') == 'file:' + urllib_request.pathname2url(path)
 
 
 @pytest.mark.parametrize("url,win_expected,non_win_expected", [
